+++ src/util/Data.java
@@ -16,7 +16,7 @@ public class Data extends java.util.Date{
     * Retorna a diferença, em segundos, entre duas datas, a primeira deve ser menor do que a segunda para que .
     * @author ASSUERO
     */
-    public static long DiferencaEmSegundos(Date data1, Date data2){
+    public static double DiferencaEmSegundos(Date data1, Date data2){
         long intDiferenca = 0;
         Calendar cldData1 = Calendar.getInstance();
         cldData1.setTime(data1);
